Update the git workflow section of the README
The original author of the workflow has updated his explanation. He has
also included a few options, and I made some choices. This new section
includes the new link, and it also includes information about which
choices I made. Now, all that needs to be done is to write a gitconfig
and scripts to manage the workflow.