Refactor so that user code has access to what it needs
This adds several new things. The biggest thing is that the user will
get handles to various things: screens (windows), areas (a portion of a
window), and area types (which define what each area draws and what
events it receives). This setup is important for the core work in Wima.