Remove the math files
Turns out that glfw already gives me the mouse coordinates in pixels,
even though they are doubles. That means that I can just cast them to
integers and be done with it. As such, I don't need the math function
that I wrote. However, it will still be in the history, just in case.