Adjust the user callbacks to take a type handle
I don't want to expose the innards of areas to the user, so I just
decided to give the user the type handle, which will allow them to get
the user pointer. However, it also means that the user will not be able
to have different data for areas of the same type. I don't think that's
much of a problem because they can just create a different area type.