Change the name of fields in the global struct for region refactor
***DOES NOT COMPILE!!!***

Because regions are now the template for areas, I decided to simplify in
another way by making workspaces templates for windows. After this
refactor, windows will just contain a tree of areas, and workspaces will
only live as templates in the global struct, just as regions only live
as templates (for areas) in the global struct. The fields in the global
struct were renamed to reflect that.