Removed some workspace code and redid the structs
***DOES NOT COMPILE!!!***

This commit changes the workspace data structures and code for the
region/workspace refactor that is happening right now. It removes code
to handle events (since workspaces are now just templates, they don't
handle events) and consolidates the WimaWksp and WimaWkspType structs
into one, since workspaces are now the "type" for windows.