Add batch event processing to windows
This change may not make sense right now, but it will make sense down
the road. The problem is that OUI, which I am going to use to do layout,
does batch event processing, and I decided that it sounded like a good
idea, especially since it will make it easier to integrate OUI in the
future.

This commit adds a couple of new types, most notably WimaEventType (an
enum) and WimaEvent (a struct that can hold any type of event data). All
of the callbacks (except the window close callback, for reasons laid out
in a comment in the code) were changed to only add to the window's list
of events. Then, at the end of the event loop, we call the function
wima_window_processEvents, which processes all of the events (calls all
user callbacks).