Add menu offsets to Wima
This makes it so that the menu pops up in the same relative location to
the mouse as it was when the user clicked on an item last time. This is
useful for Blender-like context menus that pop up options on an
operation that has been selected. Blender also does it when you save. It
asks you if you want to save over the current file, and needing to move
the mouse to click yes is annoying. Blender does not require you to, and
neither should Wima.