Make font and icon sheet global traits
This required a lot of work. It turns out that I can only make the path
global, and then for each window (and its NanoVG context), I load the
font and icon sheet. Also, it turns out that widgets will need the
NVGcontext, the font ID, and the icon sheet ID, so I made a struct to
give it to them. It was easier to just use that struct within the window
struct as well.