Add prop list to the types of props
This is so I can more easily implement themes in Wima while still
allowing the user to change them, as well as still easily generating
a user preferences screen. Since each prop type is going to be hooked
up to a specific type of UI item, it will be easy.