Fix a bug I found reading GLFW docs
The bug is that GLFW says that casting only works for positive numbers,
but when the cursor is "disabled," GLFW can return any range that
doubles can hold, which could be used for something like Blender's
wraparound cursor behavior. Thus, I changed the callback to do as GLFW
docs said: use floor().