Add icons to properties
To do this, I shortened refs, so properties can only have 65,535
children. I do not think this is a bad thing because lists can still be
drawn on the screen, and that's a big list. Also, when I added refs, I
shortened the hash to 32 bits, but I think that's enough to generally
speed things up.