Make windows fit exactly into an integer number of cache lines
This commit makes windows fit exactly into 42 cache lines (assuming a
cache line size of 64). It also pust the event and widget arrays right
onto cache lines as well. Because all the cache line sizes are even,
this even works for 128-byte cache lines.

I wanted to do this because putting the events array right on a cache
line, as well as putting the window on a cache line, has the potential
to stop a lot of cache misses when there is more than one event in the
queue.