Fail in wima_xinit() if Wima has already been initialized
This required me to write a constructor function that set the name to
NULL, since that is what is tested. Also, I just had the entire WimaG wg
struct memset to 0 at the beginning of the function. That saves lines of
code and will make sure I don't miss anything in the future. I also kept
the memset in wima_init(), not the constructor, in case clients end up
having to initialize Wima twice while running.