Fix two menu bugs
The first was that the window was being laid out when it was not
necessary. The second was that the wrong click function was happening,
or no click was registered at all. This was because I was not getting
items from the menu properly and ended up looping through the pool of
menu items instead.