Fix crash on startup
This crash was caused by overflowing a uint16_t when assigning to it
when calling wima_window_setMinSize(). The reason this was happening was
because I was doing to wrong thing to figure out the minimum size of a
parent; I was actually thinking that I had to do it in the child, and
set the minimum size of the child. But no, it had to happen in the
parent.

That wrong thinking actually led to some ugly code, too, so this commit
also removes that.