Remove menus as their own concept
*** NOT TESTED!!! ***

Menus are now contained as properties by themselves. They are collection
properties (along with three others). The way to make a menu is to make
a collection property and then put the property on the window.

This necessitated a lot of changing, because menus could now take full
advantage of layout and drawing. A lot of duplicate/ugly code was
removed.

However, this change has not been tested, and will not be tested until
Yera needs menus or overlays, which should probably be soon, considering
that I should make a menu bar at the top.

That could be one thing that Wima also provides. In fact, it *will* be
one more thing Wima provides.